drivers: broadcom: save return value of pci_find_capability() in u8
Callers of pci_find_capability should save the return value in u8.
change the type of pcix_cap from int to u8, to match the specification.

+++ drivers/net/ethernet/broadcom/tg3.h
@@ -3268,7 +3268,7 @@ struct tg3 {
 
 	int				pci_fn;
 	int				msi_cap;
-	int				pcix_cap;
+	u8				pcix_cap;
 	int				pcie_readrq;
 
 	struct mii_bus			*mdio_bus;
